notes/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
python-jose
python-dateutil
python-multipart
httpx
pydantic==1.10.15
